+++ apps/mobile/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useAuth'
 export * from './useCachedResources'
 export * from './useIsForeground'
 export * from './useIsOnline'
